+++ isodata/pjm.py
@@ -80,6 +80,7 @@ class PJM(ISOBase):
         return self._yesterday_from_historical(self.get_historical_demand)
 
     def get_historical_demand(self, date):
+        # can support a load area
         date = date = isodata.utils._handle_date(date)
         tomorrow = date + pd.DateOffset(1)
 
@@ -117,6 +118,13 @@ class PJM(ISOBase):
             "https://dataminer2.pjm.com/config/settings.json")
 
         return settings["subscriptionKey"]
+
+    # def get_historical_lmp(self, date, market: str, nodes: list):
+        # doesn't have current day
+        # REAL_TIME_5_MIN = "REAL_TIME_5_MIN"
+        # https://dataminer2.pjm.com/feed/rt_fivemin_hrl_lmps
+        # REAL_TIME_HOURLY = "REAL_TIME_HOURLY"
+        # https://dataminer2.pjm.com/feed/rt_hrl_lmps
 
 
 """